requirements.txt
asgiref==3.7.2
Django==4.2.2
django-admin-multiple-choice-list-filter==0.1.1
django-admin-rangefilter==0.10.0
django-autocomplete-light==3.9.7
django-daterangefilter==1.0.0
django-queryset-csv==1.1.0
djangorestframework==3.14.0
install==1.3.5
JPype1==1.4.1
numpy==1.25.0
packaging==23.1
Pillow==9.5.0
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
unicodecsv==0.14.1
